Revert "Move rundb code to module_utils"
