--- linchpin/provision/module_utils/linchpin_rundb.py
@@ -1,1 +1,0 @@
-../../linchpin_rundb.py
